# project/book-network-ui/src/app/services/models/registration-request.ts
/* tslint:disable */
/* eslint-disable */
export interface RegistrationRequest {
  username: string;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  confirmationPassword: string;
}